+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/features/RSFeatures.java
@@ -3,6 +3,7 @@ package com.telepathicgrunt.repurposedstructures.world.features;
 import java.util.Random;
 
 import com.telepathicgrunt.repurposedstructures.utils.RegUtil;
+import com.telepathicgrunt.repurposedstructures.world.features.structures.BadlandsTempleStructure;
 import com.telepathicgrunt.repurposedstructures.world.features.structures.JungleFortressStructure;
 import com.telepathicgrunt.repurposedstructures.world.features.structures.NetherTempleStructure;
 import com.telepathicgrunt.repurposedstructures.world.features.structures.RSMineshaftConfig;
@@ -49,6 +50,7 @@ public class RSFeatures
 	public static Structure<NoFeatureConfig>				STRONGHOLD				= new RSStrongholdStructure(NoFeatureConfig::deserialize);
 	public static Structure<NoFeatureConfig>				JUNGLE_FORTRESS			= new JungleFortressStructure(NoFeatureConfig::deserialize);
 	public static Structure<NoFeatureConfig>				NETHER_TEMPLE			= new NetherTempleStructure(NoFeatureConfig::deserialize);
+	public static Structure<NoFeatureConfig>				BADLANDS_TEMPLE			= new BadlandsTempleStructure(NoFeatureConfig::deserialize);
 
 
 	public static void registerFeatures(Register<Feature<?>> event)
@@ -78,6 +80,7 @@ public class RSFeatures
 		RegUtil.register(registry, MINESHAFT, "mineshaft");
 		RegUtil.register(registry, JUNGLE_FORTRESS, "jungle_fortress");
 		RegUtil.register(registry, NETHER_TEMPLE, "nether_temple");
+		RegUtil.register(registry, BADLANDS_TEMPLE, "badlands_temple");
 
 		STRONGHOLD.setRegistryName(new ResourceLocation("minecraft:stronghold"));
 		registry.register(STRONGHOLD);