+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/features/structures/StructurePieces.java
@@ -39,7 +39,8 @@ public class StructurePieces
 	public static IStructurePieceType	SHSTARTRS		= RSStrongholdPieces.EntranceStairs::new;
 	public static IStructurePieceType	SHSRS			= RSStrongholdPieces.Straight::new;
 	public static IStructurePieceType	SHSSDRS			= RSStrongholdPieces.StairsStraight::new;
-	public static IStructurePieceType 	NTP				= NetherTemplePiece::new;
+	public static IStructurePieceType 	NTHRTMP				= NetherTemplePiece::new;
+	public static IStructurePieceType 	BLTMP				= BadlandsTemplePiece::new;
 
 
 	public static void registerStructurePieces()
@@ -76,7 +77,8 @@ public class StructurePieces
 		RegUtil.register(SHSTARTRS, 	"SHStartRS");
 		RegUtil.register(SHSRS, 		"SHSRS");
 		RegUtil.register(SHSSDRS, 		"SHSSDRS");
-		RegUtil.register(NTP, 			"NTP");
+		RegUtil.register(NTHRTMP, 		"NthrTmP");
+		RegUtil.register(BLTMP, 		"BlTmP");
 	}
 
 }